Update Tuvok from it's repo, roughly version `0.6'.
This updates Tuvok to the bleeding edge version, which should
become 1.0 with few changes other than the DX renderer.

Updating the list of source files seemed painful.  This also
changes the makefile to just include another file with the list of
Tuvok sources.  The external file can then easily generated via a
find | awk or similar.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6587 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/third_party_builtin/tuvok/tuvok/IO/UVF/Histogram1DDataBlock.cpp
@@ -17,19 +17,19 @@ Histogram1DDataBlock::Histogram1DDataBlock(const Histogram1DDataBlock &other) :
 }
 
 Histogram1DDataBlock& Histogram1DDataBlock::operator=(const Histogram1DDataBlock& other) {
-	strBlockID = other.strBlockID;
-	ulBlockSemantics = other.ulBlockSemantics;
-	ulCompressionScheme = other.ulCompressionScheme;
-	ulOffsetToNextDataBlock = other.ulOffsetToNextDataBlock;
+  strBlockID = other.strBlockID;
+  ulBlockSemantics = other.ulBlockSemantics;
+  ulCompressionScheme = other.ulCompressionScheme;
+  ulOffsetToNextDataBlock = other.ulOffsetToNextDataBlock;
 
-	m_vHistData = other.m_vHistData;
+  m_vHistData = other.m_vHistData;
 
-	return *this;
+  return *this;
 }
 
 
 Histogram1DDataBlock::Histogram1DDataBlock(LargeRAWFile* pStreamFile, UINT64 iOffset, bool bIsBigEndian) {
-	GetHeaderFromFile(pStreamFile, iOffset, bIsBigEndian);	
+  GetHeaderFromFile(pStreamFile, iOffset, bIsBigEndian);
 }
 
 Histogram1DDataBlock::~Histogram1DDataBlock() 
@@ -37,19 +37,19 @@ Histogram1DDataBlock::~Histogram1DDataBlock()
 }
 
 DataBlock* Histogram1DDataBlock::Clone() {
-	return new Histogram1DDataBlock(*this);
+  return new Histogram1DDataBlock(*this);
 }
 
 UINT64 Histogram1DDataBlock::GetHeaderFromFile(LargeRAWFile* pStreamFile, UINT64 iOffset, bool bIsBigEndian) {
   UINT64 iStart = iOffset + DataBlock::GetHeaderFromFile(pStreamFile, iOffset, bIsBigEndian);
-	pStreamFile->SeekPos(iStart);
+  pStreamFile->SeekPos(iStart);
 
-	UINT64 ulElementCount;
-	pStreamFile->ReadData(ulElementCount, bIsBigEndian);
+  UINT64 ulElementCount;
+  pStreamFile->ReadData(ulElementCount, bIsBigEndian);
 
   m_vHistData.resize(size_t(ulElementCount));
   pStreamFile->ReadRAW((unsigned char*)&m_vHistData[0], ulElementCount*sizeof(UINT64));
-	return pStreamFile->GetPos() - iOffset;
+  return pStreamFile->GetPos() - iOffset;
 }
 
 bool Histogram1DDataBlock::Compute(RasterDataBlock* source) {
@@ -61,7 +61,7 @@ bool Histogram1DDataBlock::Compute(RasterDataBlock* source) {
   //       this should be changed to a more general approach
   vector<UINT64> vSmallestLOD = source->GetSmallestBrickIndex();
   const vector<UINT64>& vBricks = source->GetBrickCount(vSmallestLOD);
-  for (unsigned int i = 0;i<vBricks.size();i++) if (vBricks[i] != 1) return false;
+  for (size_t i = 0;i<vBricks.size();i++) if (vBricks[i] != 1) return false;
   
   // create temp histogram 
   size_t iValueRange = size_t(1<<(source->ulElementBitSize[0][0]));
@@ -110,28 +110,28 @@ bool Histogram1DDataBlock::Compute(RasterDataBlock* source) {
   for (size_t i = 0;i<iValueRange;i++) m_vHistData[i] = vTmpHist[i];
 
   // set data block information
-	strBlockID = "1D Histogram for datablock " + source->strBlockID;
+  strBlockID = "1D Histogram for datablock " + source->strBlockID;
 
   return true;
 }
 
 UINT64 Histogram1DDataBlock::CopyToFile(LargeRAWFile* pStreamFile, UINT64 iOffset, bool bIsBigEndian, bool bIsLastBlock) {
   UINT64 iStart = iOffset + DataBlock::CopyToFile(pStreamFile, iOffset, bIsBigEndian, bIsLastBlock);
-	pStreamFile->SeekPos(iStart);
+  pStreamFile->SeekPos(iStart);
 
-	UINT64 ulElementCount = UINT64(m_vHistData.size());
+  UINT64 ulElementCount = UINT64(m_vHistData.size());
   pStreamFile->WriteData(ulElementCount, bIsBigEndian);
   pStreamFile->WriteRAW((unsigned char*)&m_vHistData[0], ulElementCount*sizeof(UINT64));
 
-	return pStreamFile->GetPos() - iOffset;  
+  return pStreamFile->GetPos() - iOffset;
 }
 
 
 UINT64 Histogram1DDataBlock::GetOffsetToNextBlock() const {
-	return DataBlock::GetOffsetToNextBlock() + ComputeDataSize();
+  return DataBlock::GetOffsetToNextBlock() + ComputeDataSize();
 }
 
 UINT64 Histogram1DDataBlock::ComputeDataSize() const {
-	return sizeof(UINT64) +								  // length of the vector
-		   m_vHistData.size()*sizeof(UINT64);  // the vector itself
+  return sizeof(UINT64) +                  // length of the vector
+       m_vHistData.size()*sizeof(UINT64);  // the vector itself
 }
